+++ Tchap/Modules/Room/DataSources/RoomDataSource.m
@@ -69,6 +69,26 @@
     return self;
 }
 
+- (void)finalizeInitialization
+{
+    [super finalizeInitialization];
+
+    // Sadly, we need to make sure we have fetched all room members from the HS
+    // to be able to display read receipts
+    if (![self.mxSession.store hasLoadedAllRoomMembersForRoom:self.roomId])
+    {
+        [self.room members:^(MXRoomMembers *roomMembers) {
+            NSLog(@"[MXKRoomDataSource] finalizeRoomDataSource: All room members have been retrieved");
+
+            // Refresh the full table
+            [self.delegate dataSource:self didCellChange:nil];
+
+        } failure:^(NSError *error) {
+            NSLog(@"[MXKRoomDataSource] finalizeRoomDataSource: Cannot retrieve all room members");
+        }];
+    }
+}
+
 - (void)updateEventFormatter
 {
     // Set a new event formatter
@@ -284,7 +304,7 @@
                         if (roomMembers.count)
                         {
                             // Define the read receipts container, positioned on the right border of the bubble cell (Note the right margin 6 pts).
-                            MXKReceiptSendersContainer* avatarsContainer = [[MXKReceiptSendersContainer alloc] initWithFrame:CGRectMake(bubbleCell.frame.size.width - 156, bottomPositionY - 13, 150, 12) andRestClient:self.mxSession.matrixRestClient];
+                            MXKReceiptSendersContainer* avatarsContainer = [[MXKReceiptSendersContainer alloc] initWithFrame:CGRectMake(bubbleCell.frame.size.width - 156, bottomPositionY - 13, 150, 12) andMediaManager:self.mxSession.mediaManager];
                             
                             // Custom avatar display
                             avatarsContainer.maxDisplayedAvatars = 5;